requirements.txt
streamlit==1.25.0
paho-mqtt==1.6.1
